refactor(folder-structure): add folders to project files
